fix(amount type)

+++ models/trade/closedTrade.go
@@ -8,7 +8,7 @@ type ClosedTrade struct {
 	gorm.Model
 	UserID   uint    `gorm:"not null" json:"user_id"`
 	CryptoID uint    `gorm:"not null" json:"crypto_id"`
-	Amount   float64 `gorm:"not null" json:"amount"`
+	Amount   int 	 `gorm:"not null" json:"amount"`
 	BuyFee   int     `gorm:"not null" json:"buy_fee"`
 	SellFee  int     `gorm:"not null" json:"sell_fee"`
 	Profit   int     `gorm:"not null" json:"profit"`
